backend deployment:
Issues faced during deployment and how I solved it:

Issue 1: deploying the celery build with chrome driver by installing Google Chrome and chromedriver all in the same docker image of the celery app. The Image Google Chrome Could not be installed due to some error in dpkg and wget packages in the docker container.

Fix 1: After search, other methods of executing selenium scraping that don't involve installing chrome in the same driver was either by using browserless.io image or the standalone official selenium/chrome image. In this project I went through by using browserless.io sine it is easier to configure and it has built in support for connecting to selenium. Additionally, several options were added to avoid being detected by the indeed server as a bot.

Issue 2: Some assets in the frontend were not picked and processed during build stage.

Fix 2: A solution was to import the assists into the components script files and use data properties, or use relative path.

Issue 3: Venv and npm packges folder were uploaded and included in the final images which lead to huge image size.

Fix 3: .dockerignore file was added that ignored those files when creating and copying the folders from base to image base.

Besides that, for frontend deployment, building the project and then only uploading the build dist folder and run it using http-server was found to be better and easier solution instad of building the image and downloading all packages in the server instead.